Added clearing-methods to my_sql.py
Added "test_02_messages_of_discussion_link"-test to test_forum_collector.py

Added additional comments

+++ application/tests/A_DataCollectors/ForumCollector/test_forum_collector.py
@@ -1,4 +1,5 @@
 import unittest
+import datetime
 
 from application.A_DataCollectors.ForumCollector.forum_collector import ForumCollector
 from application.B_Database.my_sql import DatabaseManager
@@ -10,6 +11,7 @@ class TestForumCollector(unittest.TestCase):
     def setUpClass(cls):
         cls.db = DatabaseManager(user='root', password='', host='localhost', database_name='test_cassidy')
         cls.db.connect()
+        cls.db.create_tables()
         cls.category_id = cls.db.add_category('Soccer')
         cls.forum_id = cls.db.add_forum('PSV 1: Selectie & Technische Staf',
                                         'https://forum.psv.nl/index.php?forums/psv-1-selectie-technische-staf.11/',
@@ -35,6 +37,15 @@ class TestForumCollector(unittest.TestCase):
         cls.db.close()
 
     def test_01_store_discussions_of_forum_link(self):
+        # TODO: Store the forum itself in the database
+        """
+        This test consists of the following steps:
+        1. Scraping discussions from the forum using "ForumCollector.scrape_discussions_from_forum"
+        2. For each discussion:
+            2.1 Storing discussion info as a dict using "ForumCollector.return_discussion_info_from_scraped"
+            2.2 Storing the discussion in the database using "DatabaseManager.add_discussion"
+            2.3 Deleting the discussion from the database using "DatabaseManager.delete_discussion"
+        """
         discussion_class = "structItem structItem--thread js-inlineModContainer js-threadListItem"
         full_discussion_class = False
         discussion_name_class = "structItem-title"
@@ -48,9 +59,10 @@ class TestForumCollector(unittest.TestCase):
 
         for discussion in discussions:
             discussion_info = self.psv_collector.return_discussion_info_from_scraped(discussion, discussion_name_class,
-                                                                       discussion_creation_date_class,
-                                                                       discussion_views_class, discussion_replies_class,
-                                                                       discussion_last_post_time_class)
+                                                                                     discussion_creation_date_class,
+                                                                                     discussion_views_class,
+                                                                                     discussion_replies_class,
+                                                                                     discussion_last_post_time_class)
             discussion_id = self.db.add_discussion(
                 discussion_info["name"],
                 discussion_info["link"],
@@ -61,12 +73,62 @@ class TestForumCollector(unittest.TestCase):
                 discussion_info["forum_id"]
             )
 
-            self.db.delete_discussion(discussion_id)
+        self.db.clear_discussion_table()
 
         print_colored_text("Discussions stored and deleted: " + str(len(discussions)), "green")
 
     def test_02_messages_of_discussion_link(self):
-        pass
+        # TODO: Store the discussion itself in the database
+        """
+        This test consists of the following steps:
+        1. Scraping messages from the discussion using "ForumCollector.scrape_messages_from_discussion"
+        2. For each message:
+            2.1 Storing message info as a dict using "ForumCollector.return_message_info_from_scraped"
+            2.2 Storing the message in the database using "DatabaseManager.add_message"
+            2.3 Deleting the message from the database using "DatabaseManager.delete_message"
+        :return:
+        """
+        discussion_id = None
+        try:
+            discussion_link = "https://forum.psv.nl/index.php?threads/guus-til-m.1346/"
+            discussion_creation_date = datetime.datetime.now().date()
+            discussion_id = self.db.add_discussion("Guus Til (M)", discussion_link, discussion_creation_date,
+                                                   0, 0, discussion_creation_date, self.forum_id)
+            message_class = "message message--post js-post js-inlineModContainer"
+            full_message_class = False
+            message_text_class = "bbWrapper"
+            message_date_class = "u-dt"
+            message_author_class = "username"
+
+            messages = self.psv_collector.scrape_messages_from_discussion(discussion_link=discussion_link,
+                                                                          message_class=message_class,
+                                                                          full_message_class=full_message_class,
+                                                                          via_link=True)["messages"]
+            print_colored_text("Messages scraped: " + str(len(messages)), "green")
+            for message in messages:
+                message_info = self.psv_collector.return_message_info_from_scraped(message, message_text_class,
+                                                                                   message_date_class,
+                                                                                   message_author_class, discussion_id)
+
+                self.db.add_author(message_info["author"], self.psv_collector.identification)
+                user_id = self.db.select_author_by_username_and_forum_id(message_info["author"],
+                                                                         self.psv_collector.identification)["id"]
+
+                self.db.add_message(
+                    message_info["text"],
+                    message_info["date"],
+                    user_id,
+                    message_info["discussion_id"]
+                )
+
+            print_colored_text("Messages stored and deleted: " + str(len(messages)), "green")
+
+            self.db.clear_message_table()
+            self.db.clear_author_table()
+
+        finally:
+            if discussion_id:
+                self.db.delete_discussion(discussion_id)
 
     def test_03_discussions_of_forum_id(self):
         pass
